+++ lib/features/home/domain/entities/candidate.dart
@@ -48,11 +48,3 @@ class Candidate with _$Candidate {
     return fullName;
   }
 }
-
-extension IntX on int {
-  String toDateFormat([String? format]) {
-    final date = DateTime.fromMillisecondsSinceEpoch(toInt());
-    final formatter = DateFormat(format ?? 'yyyy-MM-dd');
-    return formatter.format(date);
-  }
-}
